requirements.txt
numpy~=1.26.4
scikit-learn~=1.5.1
scikit-image~=0.24.0
pillow~=10.4.0
tensorflow~=2.9.1
keras~=2.9.0
pynput~=1.7.7
h5py~=3.11.0
keyboard~=0.13.5
pygetwindow~=0.0.9
